+++ test/body-parser.test.ts
@@ -5,9 +5,7 @@ import { describe, expect, it } from 'bun:test'
 describe('Body Parser', () => {
 	it('handle onParse', async () => {
 		const app = new KingWorld()
-			.onParse(async (request) => {
-				const contentType = request.headers.get('content-type') ?? ''
-
+			.onParse((request, contentType) => {
 				switch (contentType) {
 					case 'application/kingworld':
 						return request.text()
@@ -32,9 +30,7 @@ describe('Body Parser', () => {
 
 	it("handle .on('parse')", async () => {
 		const app = new KingWorld()
-			.on('parse', async (request) => {
-				const contentType = request.headers.get('content-type') ?? ''
-
+			.on('parse', (request, contentType) => {
 				switch (contentType) {
 					case 'application/kingworld':
 						return request.text()
@@ -55,5 +51,30 @@ describe('Body Parser', () => {
 		)
 
 		expect(await res.text()).toBe(':D')
+	})
+
+	it('overwrite default parser', async () => {
+		const app = new KingWorld()
+			.onParse((request, contentType) => {
+				switch (contentType) {
+					case 'text/plain':
+						return 'Overwrited'
+				}
+			})
+			.post('/', ({ body }) => body)
+			.listen(3000)
+
+		const res = await app.handle(
+			new Request('/', {
+				method: 'POST',
+				body: ':D',
+				headers: {
+					'content-type': 'text/plain',
+					'content-length': '2'
+				}
+			})
+		)
+
+		expect(await res.text()).toBe('Overwrited')
 	})
 })